+++ docs/changes.rst
@@ -8,6 +8,7 @@ Released on ? 2018.
 
 * Increase required translate-toolkit version to 2.3.1.
 * Add Netease Sight translate service.
+* Add support for Pagure webhooks.
 
 weblate 3.2.1
 -------------